+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-0.4dev
-------
+0.4
+---
 
 - Implement ``bda.plone.orders.interfaces.IVendor`` for AT and DX Based
   content.